+++ src/anomalib/metrics/per_image/_validate.py
@@ -11,8 +11,22 @@ https://github.com/openvinotoolkit/anomalib/issues/2093
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
+import logging
+from collections import OrderedDict
+from collections.abc import Sequence
+from typing import TYPE_CHECKING
+
 import torch
 from torch import Tensor
+
+from anomalib.data.utils.path import validate_path
+
+from .utils import images_classes_from_masks
+
+if TYPE_CHECKING:
+    from .pimo import AUPIMOResult
+
+logger = logging.getLogger(__name__)
 
 
 def is_num_threshs_gte2(num_threshs: int) -> None:
@@ -347,3 +361,323 @@ def is_per_image_rate_curves(rate_curves: Tensor, nan_allowed: bool, decreasing:
             "but got non-monotonically increasing values."
         )
         raise ValueError(msg)
+
+
+def is_scores_batch(scores_batch: torch.Tensor) -> None:
+    """scores_batch (torch.Tensor): floating (N, D)."""
+    if not isinstance(scores_batch, torch.Tensor):
+        msg = f"Expected `scores_batch` to be an torch.Tensor, but got {type(scores_batch)}"
+        raise TypeError(msg)
+
+    if not scores_batch.dtype.is_floating_point:
+        msg = (
+            "Expected `scores_batch` to be an floating torch.Tensor with anomaly scores_batch,"
+            f" but got torch.Tensor with dtype {scores_batch.dtype}"
+        )
+        raise TypeError(msg)
+
+    if scores_batch.ndim != 2:
+        msg = f"Expected `scores_batch` to be 2D, but got {scores_batch.ndim}"
+        raise ValueError(msg)
+
+
+def is_gts_batch(gts_batch: torch.Tensor) -> None:
+    """gts_batch (torch.Tensor): boolean (N, D)."""
+    if not isinstance(gts_batch, torch.Tensor):
+        msg = f"Expected `gts_batch` to be an torch.Tensor, but got {type(gts_batch)}"
+        raise TypeError(msg)
+
+    if gts_batch.dtype != torch.bool:
+        msg = (
+            "Expected `gts_batch` to be an boolean torch.Tensor with anomaly scores_batch,"
+            f" but got torch.Tensor with dtype {gts_batch.dtype}"
+        )
+        raise TypeError(msg)
+
+    if gts_batch.ndim != 2:
+        msg = f"Expected `gts_batch` to be 2D, but got {gts_batch.ndim}"
+        raise ValueError(msg)
+
+
+def has_at_least_one_anomalous_image(masks: torch.Tensor) -> None:
+    is_masks(masks)
+    image_classes = images_classes_from_masks(masks)
+    if (image_classes == 1).sum() == 0:
+        msg = "Expected at least one ANOMALOUS image, but found none."
+        raise ValueError(msg)
+
+
+def has_at_least_one_normal_image(masks: torch.Tensor) -> None:
+    is_masks(masks)
+    image_classes = images_classes_from_masks(masks)
+    if (image_classes == 0).sum() == 0:
+        msg = "Expected at least one NORMAL image, but found none."
+        raise ValueError(msg)
+
+
+def joint_validate_threshs_shared_fpr(threshs: torch.Tensor, shared_fpr: torch.Tensor) -> None:
+    if threshs.shape[0] != shared_fpr.shape[0]:
+        msg = (
+            "Expected `threshs` and `shared_fpr` to have the same number of elements, "
+            f"but got {threshs.shape[0]} != {shared_fpr.shape[0]}"
+        )
+        raise ValueError(msg)
+
+
+def is_per_image_tprs(per_image_tprs: torch.Tensor, image_classes: torch.Tensor) -> None:
+    is_images_classes(image_classes)
+    # general validations
+    is_per_image_rate_curves(
+        per_image_tprs,
+        nan_allowed=True,  # normal images have NaN TPRs
+        decreasing=None,  # not checked here
+    )
+
+    # specific to anomalous images
+    is_per_image_rate_curves(
+        per_image_tprs[image_classes == 1],
+        nan_allowed=False,
+        decreasing=True,
+    )
+
+    # specific to normal images
+    normal_images_tprs = per_image_tprs[image_classes == 0]
+    if not normal_images_tprs.isnan().all():
+        msg = "Expected all normal images to have NaN TPRs, but some have non-NaN values."
+        raise ValueError(msg)
+
+
+def is_source_images_paths(paths: Sequence[str], expected_num_paths: int | None) -> None:
+    if not isinstance(paths, list):
+        msg = f"Expected paths to be a list, but got {type(paths)}."
+        raise TypeError(msg)
+
+    for idx, path in enumerate(paths):
+        try:
+            msg = f"Invalid path at index {idx}: {path}"
+            validate_path(
+                path,
+                # not necessary to exist because the metric can be computed
+                # directly from the anomaly maps and masks, without the images
+                should_exist=False,
+            )
+
+        except TypeError as ex:
+            raise TypeError(msg) from ex
+
+        except ValueError as ex:
+            raise ValueError(msg) from ex
+
+        if not isinstance(path, str):
+            # this will eventually be serialized to a file, so we don't want pathlib objects keep it simple
+            msg = f"Expected path to be a string, but got {type(path)}."
+            raise TypeError(msg)
+
+    if expected_num_paths is None:
+        return
+
+    if len(paths) != expected_num_paths:
+        msg = f"Invalid `paths` argument. Expected {expected_num_paths} paths, but got {len(paths)} instead."
+        raise ValueError(msg)
+
+
+def is_per_image_scores(per_image_scores: torch.Tensor) -> None:
+    if per_image_scores.ndim != 1:
+        msg = f"Expected per-image scores to be 1D, but got {per_image_scores.ndim}D."
+        raise ValueError(msg)
+
+
+def is_image_class(image_class: int) -> None:
+    if image_class not in {0, 1}:
+        msg = f"Expected image class to be either 0 for 'normal' or 1 for 'anomalous', but got {image_class}."
+        raise ValueError(msg)
+
+
+def is_models_ordered(models_ordered: tuple[str, ...]) -> None:
+    if not isinstance(models_ordered, tuple):
+        msg = f"Expected models ordered to be a tuple, but got {type(models_ordered)}."
+        raise TypeError(msg)
+
+    if len(models_ordered) < 2:
+        msg = f"Expected models ordered to have at least 2 models, but got {len(models_ordered)}."
+        raise ValueError(msg)
+
+    for model_name in models_ordered:
+        if not isinstance(model_name, str):
+            msg = f"Expected model name to be a string, but got {type(model_name)} for model {model_name}."
+            raise TypeError(msg)
+
+        if model_name == "":
+            msg = "Expected model name to be non-empty, but got empty string."
+            raise ValueError(msg)
+
+    num_redundant_models = len(models_ordered) - len(set(models_ordered))
+    if num_redundant_models > 0:
+        msg = f"Expected models ordered to have unique models, but got {num_redundant_models} redundant models."
+        raise ValueError(msg)
+
+
+def is_confidences(confidences: dict[tuple[str, str], float]) -> None:
+    if not isinstance(confidences, dict):
+        msg = f"Expected confidences to be a dict, but got {type(confidences)}."
+        raise TypeError(msg)
+
+    for (model1, model2), confidence in confidences.items():
+        if not isinstance(model1, str):
+            msg = f"Expected model name to be a string, but got {type(model1)} for model {model1}."
+            raise TypeError(msg)
+
+        if not isinstance(model2, str):
+            msg = f"Expected model name to be a string, but got {type(model2)} for model {model2}."
+            raise TypeError(msg)
+
+        if not isinstance(confidence, float):
+            msg = f"Expected confidence to be a float, but got {type(confidence)} for models {model1} and {model2}."
+            raise TypeError(msg)
+
+        if not (0 <= confidence <= 1):
+            msg = f"Expected confidence to be between 0 and 1, but got {confidence} for models {model1} and {model2}."
+            raise ValueError(msg)
+
+
+def joint_validate_models_ordered_and_confidences(
+    models_ordered: tuple[str, ...],
+    confidences: dict[tuple[str, str], float],
+) -> None:
+    num_models = len(models_ordered)
+    expected_num_pairs = num_models * (num_models - 1)
+
+    if len(confidences) != expected_num_pairs:
+        msg = f"Expected {expected_num_pairs} pairs of models, but got {len(confidences)} pairs of models."
+        raise ValueError(msg)
+
+    models_in_confidences = {model for pair_models in confidences for model in pair_models}
+
+    diff = set(models_ordered).symmetric_difference(models_in_confidences)
+    if len(diff) > 0:
+        msg = (
+            "Expected models in confidences to be the same as models ordered, but got models missing in one"
+            f"of them: {diff}."
+        )
+        raise ValueError(msg)
+
+
+def is_scores_per_model_tensor(scores_per_model: dict[str, Tensor] | OrderedDict[str, Tensor]) -> None:
+    first_key_value = None
+
+    for model_name, scores in scores_per_model.items():
+        if scores.ndim != 1:
+            msg = f"Expected scores to be 1D, but got {scores.ndim}D for model {model_name}."
+            raise ValueError(msg)
+
+        num_valid_scores = scores[~torch.isnan(scores)].numel()
+
+        if num_valid_scores < 1:
+            msg = f"Expected at least 1 non-nan score, but got {num_valid_scores} for model {model_name}."
+            raise ValueError(msg)
+
+        if first_key_value is None:
+            first_key_value = (model_name, scores)
+            continue
+
+        first_model_name, first_scores = first_key_value
+
+        # same shape
+        if scores.shape[0] != first_scores.shape[0]:
+            msg = (
+                "Expected scores to have the same number of scores, "
+                f"but got ({model_name}) {scores.shape[0]} != {first_scores.shape[0]} ({first_model_name})."
+            )
+            raise ValueError(msg)
+
+        # `nan` at the same indices
+        if (torch.isnan(scores) != torch.isnan(first_scores)).any():
+            msg = (
+                "Expected `nan` values, if any, to be at the same indices, "
+                f"but there are differences between models {model_name} and {first_model_name}."
+            )
+            raise ValueError(msg)
+
+
+def is_scores_per_model_aupimoresult(
+    scores_per_model: dict[str, "AUPIMOResult"] | OrderedDict[str, "AUPIMOResult"],
+    missing_paths_ok: bool,
+) -> None:
+    first_key_value = None
+
+    for model_name, aupimoresult in scores_per_model.items():
+        if first_key_value is None:
+            first_key_value = (model_name, aupimoresult)
+            continue
+
+        first_model_name, first_aupimoresult = first_key_value
+
+        if aupimoresult.fpr_bounds != first_aupimoresult.fpr_bounds:
+            msg = (
+                "Expected AUPIMOResult objects in scores per model to have the same FPR bounds, "
+                f"but got ({model_name}) {aupimoresult.fpr_bounds} != "
+                f"{first_aupimoresult.fpr_bounds} ({first_model_name})."
+            )
+            raise ValueError(msg)
+
+    available_paths = [tuple(scores.paths) for scores in scores_per_model.values() if scores.paths is not None]
+
+    if len(set(available_paths)) > 1:
+        msg = (
+            "Expected AUPIMOResult objects in scores per model to have the same paths, "
+            "but got different paths for different models."
+        )
+        raise ValueError(msg)
+
+    if len(available_paths) != len(scores_per_model):
+        msg = "Some models have paths, while others are missing them."
+        if not missing_paths_ok:
+            raise ValueError(msg)
+        logger.warning(msg)
+
+
+def is_scores_per_model(
+    scores_per_model: dict[str, Tensor]
+    | OrderedDict[str, Tensor]
+    | dict[str, "AUPIMOResult"]
+    | OrderedDict[str, "AUPIMOResult"],
+) -> None:
+    # it has to be imported here to avoid circular imports
+    from .pimo import AUPIMOResult
+
+    if not isinstance(scores_per_model, dict | OrderedDict):
+        msg = f"Expected scores per model to be a dictionary or ordered dictionary, but got {type(scores_per_model)}."
+        raise TypeError(msg)
+
+    if len(scores_per_model) < 2:
+        msg = f"Expected scores per model to have at least 2 models, but got {len(scores_per_model)}."
+        raise ValueError(msg)
+
+    if not all(isinstance(model_name, str) for model_name in scores_per_model):
+        msg = "Expected scores per model to have model names (strings) as keys."
+        raise TypeError(msg)
+
+    first_instance = next(iter(scores_per_model.values()))
+
+    if (
+        isinstance(first_instance, Tensor)
+        and any(not isinstance(scores, Tensor) for scores in scores_per_model.values())
+    ) or (
+        isinstance(first_instance, AUPIMOResult)
+        and any(not isinstance(scores, AUPIMOResult) for scores in scores_per_model.values())
+    ):
+        msg = (
+            "Values in the scores per model dict must have the same type for values (Tensor or AUPIMOResult), "
+            "but more than one type was found."
+        )
+        raise TypeError(msg)
+
+    if isinstance(first_instance, Tensor):
+        is_scores_per_model_tensor(scores_per_model)
+        return
+
+    is_scores_per_model_tensor(
+        {model_name: scores.aupimos for model_name, scores in scores_per_model.items()},
+    )
+
+    is_scores_per_model_aupimoresult(scores_per_model, missing_paths_ok=True)
