+++ src/anomalib/metrics/per_image/pimo.py
@@ -40,456 +40,14 @@ so often times the Tensor arguments will be converted to ndarray and then valida
 # SPDX-License-Identifier: Apache-2.0
 
 import logging
-from collections.abc import Sequence
-from dataclasses import dataclass, field
 
 import torch
 from torchmetrics import Metric
 
-from anomalib.data.utils.path import validate_path
-
 from . import _validate, functional
+from .dataclasses import AUPIMOResult, PIMOResult
 
 logger = logging.getLogger(__name__)
-
-
-def _images_classes_from_masks(masks: torch.Tensor) -> torch.Tensor:
-    masks = torch.concat(masks, dim=0)
-    device = masks.device
-    image_classes = functional._images_classes_from_masks(masks)  # noqa: SLF001
-    return torch.from_numpy(image_classes, device=device)
-
-
-# =========================================== ARGS VALIDATION ===========================================
-
-
-def _validate_is_shared_fpr(shared_fpr: torch.Tensor, nan_allowed: bool = False, decreasing: bool = True) -> None:
-    _validate.is_rate_curve(shared_fpr, nan_allowed=nan_allowed, decreasing=decreasing)
-
-
-def _validate_is_image_classes(image_classes: torch.Tensor) -> None:
-    _validate.is_images_classes(image_classes)
-
-
-def _validate_is_per_image_tprs(per_image_tprs: torch.Tensor, image_classes: torch.Tensor) -> None:
-    _validate_is_image_classes(image_classes)
-    # general validations
-    _validate.is_per_image_rate_curves(
-        per_image_tprs,
-        nan_allowed=True,  # normal images have NaN TPRs
-        decreasing=None,  # not checked here
-    )
-
-    # specific to anomalous images
-    _validate.is_per_image_rate_curves(
-        per_image_tprs[image_classes == 1],
-        nan_allowed=False,
-        decreasing=True,
-    )
-
-    # specific to normal images
-    normal_images_tprs = per_image_tprs[image_classes == 0]
-    if not normal_images_tprs.isnan().all():
-        msg = "Expected all normal images to have NaN TPRs, but some have non-NaN values."
-        raise ValueError(msg)
-
-
-def _validate_is_aupimos(aupimos: torch.Tensor) -> None:
-    _validate.is_rates(aupimos, nan_allowed=True)
-
-
-def _validate_is_source_images_paths(paths: Sequence[str], expected_num_paths: int | None) -> None:
-    if not isinstance(paths, list):
-        msg = f"Expected paths to be a list, but got {type(paths)}."
-        raise TypeError(msg)
-
-    for idx, path in enumerate(paths):
-        try:
-            msg = f"Invalid path at index {idx}: {path}"
-            validate_path(
-                path,
-                # not necessary to exist because the metric can be computed
-                # directly from the anomaly maps and masks, without the images
-                should_exist=False,
-            )
-
-        except TypeError as ex:
-            raise TypeError(msg) from ex
-
-        except ValueError as ex:
-            raise ValueError(msg) from ex
-
-        if not isinstance(path, str):
-            # this will eventually be serialized to a file, so we don't want pathlib objects keep it simple
-            msg = f"Expected path to be a string, but got {type(path)}."
-            raise TypeError(msg)
-
-    if expected_num_paths is None:
-        return
-
-    if len(paths) != expected_num_paths:
-        msg = f"Invalid `paths` argument. Expected {expected_num_paths} paths, but got {len(paths)} instead."
-        raise ValueError(msg)
-
-
-# =========================================== RESULT OBJECT ===========================================
-
-
-@dataclass
-class PIMOResult:
-    """Per-Image Overlap (PIMO, pronounced pee-mo) curve.
-
-    This interface gathers the PIMO curve data and metadata and provides several utility methods.
-
-    Notation:
-        - N: number of images
-        - K: number of thresholds
-        - FPR: False Positive Rate
-        - TPR: True Positive Rate
-
-    Attributes:
-        threshs (torch.Tensor): sequence of K (monotonically increasing) thresholds used to compute the PIMO curve
-        shared_fpr (torch.Tensor): K values of the shared FPR metric at the corresponding thresholds
-        per_image_tprs (torch.Tensor): for each of the N images, the K values of in-image TPR at the corresponding
-            thresholds
-        paths (list[str]) (optional): [metadata] paths to the source images to which the PIMO curves correspond
-    """
-
-    # data
-    threshs: torch.Tensor = field(repr=False)  # shape => (K,)
-    shared_fpr: torch.Tensor = field(repr=False)  # shape => (K,)
-    per_image_tprs: torch.Tensor = field(repr=False)  # shape => (N, K)
-
-    # optional metadata
-    paths: list[str] | None = field(repr=False, default=None)
-
-    @property
-    def num_threshs(self) -> int:
-        """Number of thresholds."""
-        return self.threshs.shape[0]
-
-    @property
-    def num_images(self) -> int:
-        """Number of images."""
-        return self.per_image_tprs.shape[0]
-
-    @property
-    def image_classes(self) -> torch.Tensor:
-        """Image classes (0: normal, 1: anomalous).
-
-        Deduced from the per-image TPRs.
-        If any TPR value is not NaN, the image is considered anomalous.
-        """
-        return (~torch.isnan(self.per_image_tprs)).any(dim=1).to(torch.int32)
-
-    def __post_init__(self) -> None:
-        """Validate the inputs for the result object are consistent."""
-        try:
-            _validate.is_threshs(self.threshs)
-            _validate_is_shared_fpr(self.shared_fpr, nan_allowed=False)
-            _validate_is_per_image_tprs(self.per_image_tprs, self.image_classes)
-
-            if self.paths is not None:
-                _validate_is_source_images_paths(self.paths, expected_num_paths=self.per_image_tprs.shape[0])
-
-        except (TypeError, ValueError) as ex:
-            msg = f"Invalid inputs for {self.__class__.__name__} object. Cause: {ex}."
-            raise TypeError(msg) from ex
-
-        if self.threshs.shape != self.shared_fpr.shape:
-            msg = (
-                f"Invalid {self.__class__.__name__} object. Attributes have inconsistent shapes: "
-                f"{self.threshs.shape=} != {self.shared_fpr.shape=}."
-            )
-            raise TypeError(msg)
-
-        if self.threshs.shape[0] != self.per_image_tprs.shape[1]:
-            msg = (
-                f"Invalid {self.__class__.__name__} object. Attributes have inconsistent shapes: "
-                f"{self.threshs.shape[0]=} != {self.per_image_tprs.shape[1]=}."
-            )
-            raise TypeError(msg)
-
-    def thresh_at(self, fpr_level: float) -> tuple[int, float, float]:
-        """Return the threshold at the given shared FPR.
-
-        See `anomalib.metrics.per_image.pimo_numpy.thresh_at_shared_fpr_level` for details.
-
-        Args:
-            fpr_level (float): shared FPR level
-
-        Returns:
-            tuple[int, float, float]:
-                [0] index of the threshold
-                [1] threshold
-                [2] the actual shared FPR value at the returned threshold
-        """
-        return functional.thresh_at_shared_fpr_level(
-            self.threshs,
-            self.shared_fpr,
-            fpr_level,
-        )
-
-
-@dataclass
-class AUPIMOResult:
-    """Area Under the Per-Image Overlap (AUPIMO, pronounced a-u-pee-mo) curve.
-
-    This interface gathers the AUPIMO data and metadata and provides several utility methods.
-
-    Attributes:
-        fpr_lower_bound (float): [metadata] LOWER bound of the FPR integration range
-        fpr_upper_bound (float): [metadata] UPPER bound of the FPR integration range
-        num_threshs (int): [metadata] number of thresholds used to effectively compute AUPIMO;
-                            should not be confused with the number of thresholds used to compute the PIMO curve
-        thresh_lower_bound (float): LOWER threshold bound --> corresponds to the UPPER FPR bound
-        thresh_upper_bound (float): UPPER threshold bound --> corresponds to the LOWER FPR bound
-        aupimos (torch.Tensor): values of AUPIMO scores (1 per image)
-    """
-
-    # metadata
-    fpr_lower_bound: float
-    fpr_upper_bound: float
-    num_threshs: int
-
-    # data
-    thresh_lower_bound: float = field(repr=False)
-    thresh_upper_bound: float = field(repr=False)
-    aupimos: torch.Tensor = field(repr=False)  # shape => (N,)
-
-    # optional metadata
-    paths: list[str] | None = field(repr=False, default=None)
-
-    @property
-    def num_images(self) -> int:
-        """Number of images."""
-        return self.aupimos.shape[0]
-
-    @property
-    def num_normal_images(self) -> int:
-        """Number of normal images."""
-        return int((self.image_classes == 0).sum())
-
-    @property
-    def num_anomalous_images(self) -> int:
-        """Number of anomalous images."""
-        return int((self.image_classes == 1).sum())
-
-    @property
-    def image_classes(self) -> torch.Tensor:
-        """Image classes (0: normal, 1: anomalous)."""
-        # if an instance has `nan` aupimo it's because it's a normal image
-        return self.aupimos.isnan().to(torch.int32)
-
-    @property
-    def fpr_bounds(self) -> tuple[float, float]:
-        """Lower and upper bounds of the FPR integration range."""
-        return self.fpr_lower_bound, self.fpr_upper_bound
-
-    @property
-    def thresh_bounds(self) -> tuple[float, float]:
-        """Lower and upper bounds of the threshold integration range.
-
-        Recall: they correspond to the FPR bounds in reverse order.
-        I.e.:
-            fpr_lower_bound --> thresh_upper_bound
-            fpr_upper_bound --> thresh_lower_bound
-        """
-        return self.thresh_lower_bound, self.thresh_upper_bound
-
-    def __post_init__(self) -> None:
-        """Validate the inputs for the result object are consistent."""
-        try:
-            _validate.is_rate_range((self.fpr_lower_bound, self.fpr_upper_bound))
-            # TODO(jpcbertoldo): warn when it's too low (use parameters from the numpy code)  # noqa: TD003
-            _validate.is_num_threshs_gte2(self.num_threshs)
-            _validate_is_aupimos(self.aupimos)
-            _validate.is_thresh_bounds((self.thresh_lower_bound, self.thresh_upper_bound))
-
-            if self.paths is not None:
-                _validate_is_source_images_paths(self.paths, expected_num_paths=self.aupimos.shape[0])
-
-        except (TypeError, ValueError) as ex:
-            msg = f"Invalid inputs for {self.__class__.__name__} object. Cause: {ex}."
-            raise TypeError(msg) from ex
-
-    @classmethod
-    def from_pimoresult(
-        cls: type["AUPIMOResult"],
-        pimoresult: PIMOResult,
-        fpr_bounds: tuple[float, float],
-        num_threshs_auc: int,
-        aupimos: torch.Tensor,
-        paths: list[str] | None = None,
-    ) -> "AUPIMOResult":
-        """Return an AUPIMO result object from a PIMO result object.
-
-        Args:
-            pimoresult: PIMO result object
-            fpr_bounds: lower and upper bounds of the FPR integration range
-            num_threshs_auc: number of thresholds used to effectively compute AUPIMO;
-                         NOT the number of thresholds used to compute the PIMO curve!
-            aupimos: AUPIMO scores
-            paths: paths to the source images to which the AUPIMO scores correspond.
-        """
-        if pimoresult.per_image_tprs.shape[0] != aupimos.shape[0]:
-            msg = (
-                f"Invalid {cls.__name__} object. Attributes have inconsistent shapes: "
-                f"there are {pimoresult.per_image_tprs.shape[0]} PIMO curves but {aupimos.shape[0]} AUPIMO scores."
-            )
-            raise TypeError(msg)
-
-        if not torch.isnan(aupimos[pimoresult.image_classes == 0]).all():
-            msg = "Expected all normal images to have NaN AUPIMOs, but some have non-NaN values."
-            raise TypeError(msg)
-
-        if torch.isnan(aupimos[pimoresult.image_classes == 1]).any():
-            msg = "Expected all anomalous images to have valid AUPIMOs (not nan), but some have NaN values."
-            raise TypeError(msg)
-
-        if pimoresult.paths is not None:
-            paths = pimoresult.paths
-
-        elif paths is not None:
-            _validate_is_source_images_paths(paths, expected_num_paths=pimoresult.num_images)
-
-        fpr_lower_bound, fpr_upper_bound = fpr_bounds
-        # recall: fpr upper/lower bounds are the same as the thresh lower/upper bounds
-        _, thresh_lower_bound, __ = pimoresult.thresh_at(fpr_upper_bound)
-        _, thresh_upper_bound, __ = pimoresult.thresh_at(fpr_lower_bound)
-        # `_` is the threshold's index, `__` is the actual fpr value
-        return cls(
-            fpr_lower_bound=fpr_lower_bound,
-            fpr_upper_bound=fpr_upper_bound,
-            num_threshs=num_threshs_auc,
-            thresh_lower_bound=float(thresh_lower_bound),
-            thresh_upper_bound=float(thresh_upper_bound),
-            aupimos=aupimos,
-            paths=paths,
-        )
-
-
-# =========================================== FUNCTIONAL ===========================================
-def pimo_curves(
-    anomaly_maps: torch.Tensor,
-    masks: torch.Tensor,
-    num_threshs: int,
-    paths: list[str] | None = None,
-) -> PIMOResult:
-    """Compute the Per-IMage Overlap (PIMO, pronounced pee-mo) curves.
-
-    This torch interface is a wrapper around the numpy code.
-    The tensors are converted to numpy arrays and then passed and validated in the numpy code.
-    The results are converted back to tensors and wrapped in an dataclass object.
-
-    PIMO is a curve of True Positive Rate (TPR) values on each image across multiple anomaly score thresholds.
-    The anomaly score thresholds are indexed by a (cross-image shared) value of False Positive Rate (FPR) measure on
-    the normal images.
-
-    Details: `anomalib.metrics.per_image.pimo`.
-
-    Args' notation:
-        N: number of images
-        H: image height
-        W: image width
-        K: number of thresholds
-
-    Args:
-        anomaly_maps: floating point anomaly score maps of shape (N, H, W)
-        masks: binary (bool or int) ground truth masks of shape (N, H, W)
-        num_threshs: number of thresholds to compute (K)
-        paths: paths to the source images to which the PIMO curves correspond. Default: None.
-
-    Returns:
-        PIMOResult: PIMO curves dataclass object. See `PIMOResult` for details.
-    """
-    if paths is not None:
-        _validate_is_source_images_paths(paths, expected_num_paths=anomaly_maps.shape[0])
-
-    # other validations are done in the numpy code
-    threshs, shared_fpr, per_image_tprs, _ = functional.pimo_curves(
-        anomaly_maps,
-        masks,
-        num_threshs,
-    )
-    # _ is `image_classes` -- not needed here because it's a property in the result object
-
-    return PIMOResult(
-        threshs=threshs,
-        shared_fpr=shared_fpr,
-        per_image_tprs=per_image_tprs,
-        paths=paths,
-    )
-
-
-def aupimo_scores(
-    anomaly_maps: torch.Tensor,
-    masks: torch.Tensor,
-    num_threshs: int = 300_000,
-    fpr_bounds: tuple[float, float] = (1e-5, 1e-4),
-    force: bool = False,
-    paths: list[str] | None = None,
-) -> tuple[PIMOResult, AUPIMOResult]:
-    """Compute the PIMO curves and their Area Under the Curve (i.e. AUPIMO) scores.
-
-    This torch interface is a wrapper around the numpy code.
-    The tensors are converted to numpy arrays and then passed and validated in the numpy code.
-    The results are converted back to tensors and wrapped in an dataclass object.
-
-    Scores are computed from the integration of the PIMO curves within the given FPR bounds, then normalized to [0, 1].
-    It can be thought of as the average TPR of the PIMO curves within the given FPR bounds.
-
-    Details: `anomalib.metrics.per_image.pimo`.
-
-    Args' notation:
-        N: number of images
-        H: image height
-        W: image width
-        K: number of thresholds
-
-    Args:
-        anomaly_maps: floating point anomaly score maps of shape (N, H, W)
-        masks: binary (bool or int) ground truth masks of shape (N, H, W)
-        num_threshs: number of thresholds to compute (K)
-        fpr_bounds: lower and upper bounds of the FPR integration range
-        force: whether to force the computation despite bad conditions
-        paths: paths to the source images to which the AUPIMO scores correspond.
-
-    Returns:
-        tuple[PIMOResult, AUPIMOResult]: PIMO and AUPIMO results dataclass objects. See `PIMOResult` and `AUPIMOResult`.
-    """
-    if paths is not None:
-        _validate_is_source_images_paths(paths, expected_num_paths=anomaly_maps.shape[0])
-
-    # other validations are done in the numpy code
-
-    threshs, shared_fpr, per_image_tprs, _, aupimos, num_threshs_auc = functional.aupimo_scores(
-        anomaly_maps,
-        masks,
-        num_threshs,
-        fpr_bounds=fpr_bounds,
-        force=force,
-    )
-
-    pimoresult = PIMOResult(
-        threshs=threshs,
-        shared_fpr=shared_fpr,
-        per_image_tprs=per_image_tprs,
-        paths=paths,
-    )
-    aupimoresult = AUPIMOResult.from_pimoresult(
-        pimoresult,
-        fpr_bounds=fpr_bounds,
-        # not `num_threshs`!
-        # `num_threshs` is the number of thresholds used to compute the PIMO curve
-        # this is the number of thresholds used to compute the AUPIMO integral
-        num_threshs_auc=num_threshs_auc,
-        aupimos=aupimos,
-    )
-    return pimoresult, aupimoresult
-
-
-# =========================================== TORCHMETRICS ===========================================
 
 
 class PIMO(Metric):
@@ -546,7 +104,7 @@ class PIMO(Metric):
     @property
     def image_classes(self) -> torch.Tensor:
         """Image classes (0: normal, 1: anomalous)."""
-        return _images_classes_from_masks(self.masks)
+        return functional.images_classes_from_masks(self.masks)
 
     def __init__(
         self,
@@ -715,3 +273,122 @@ class AUPIMO(PIMO):
             is_nan = torch.isnan(aupimoresult.aupimos)
             return aupimoresult.aupimos[~is_nan].mean()
         return pimoresult, aupimoresult
+
+
+def pimo_curves(
+    anomaly_maps: torch.Tensor,
+    masks: torch.Tensor,
+    num_threshs: int,
+    paths: list[str] | None = None,
+) -> PIMOResult:
+    """Compute the Per-IMage Overlap (PIMO, pronounced pee-mo) curves.
+
+    This torch interface is a wrapper around the numpy code.
+    The tensors are converted to numpy arrays and then passed and validated in the numpy code.
+    The results are converted back to tensors and wrapped in an dataclass object.
+
+    PIMO is a curve of True Positive Rate (TPR) values on each image across multiple anomaly score thresholds.
+    The anomaly score thresholds are indexed by a (cross-image shared) value of False Positive Rate (FPR) measure on
+    the normal images.
+
+    Details: `anomalib.metrics.per_image.pimo`.
+
+    Args' notation:
+        N: number of images
+        H: image height
+        W: image width
+        K: number of thresholds
+
+    Args:
+        anomaly_maps: floating point anomaly score maps of shape (N, H, W)
+        masks: binary (bool or int) ground truth masks of shape (N, H, W)
+        num_threshs: number of thresholds to compute (K)
+        paths: paths to the source images to which the PIMO curves correspond. Default: None.
+
+    Returns:
+        PIMOResult: PIMO curves dataclass object. See `PIMOResult` for details.
+    """
+    if paths is not None:
+        _validate.is_source_images_paths(paths, expected_num_paths=anomaly_maps.shape[0])
+
+    # other validations are done in the numpy code
+    threshs, shared_fpr, per_image_tprs, _ = functional.pimo_curves(
+        anomaly_maps,
+        masks,
+        num_threshs,
+    )
+    # _ is `image_classes` -- not needed here because it's a property in the result object
+
+    return PIMOResult(
+        threshs=threshs,
+        shared_fpr=shared_fpr,
+        per_image_tprs=per_image_tprs,
+        paths=paths,
+    )
+
+
+def aupimo_scores(
+    anomaly_maps: torch.Tensor,
+    masks: torch.Tensor,
+    num_threshs: int = 300_000,
+    fpr_bounds: tuple[float, float] = (1e-5, 1e-4),
+    force: bool = False,
+    paths: list[str] | None = None,
+) -> tuple[PIMOResult, AUPIMOResult]:
+    """Compute the PIMO curves and their Area Under the Curve (i.e. AUPIMO) scores.
+
+    This torch interface is a wrapper around the numpy code.
+    The tensors are converted to numpy arrays and then passed and validated in the numpy code.
+    The results are converted back to tensors and wrapped in an dataclass object.
+
+    Scores are computed from the integration of the PIMO curves within the given FPR bounds, then normalized to [0, 1].
+    It can be thought of as the average TPR of the PIMO curves within the given FPR bounds.
+
+    Details: `anomalib.metrics.per_image.pimo`.
+
+    Args' notation:
+        N: number of images
+        H: image height
+        W: image width
+        K: number of thresholds
+
+    Args:
+        anomaly_maps: floating point anomaly score maps of shape (N, H, W)
+        masks: binary (bool or int) ground truth masks of shape (N, H, W)
+        num_threshs: number of thresholds to compute (K)
+        fpr_bounds: lower and upper bounds of the FPR integration range
+        force: whether to force the computation despite bad conditions
+        paths: paths to the source images to which the AUPIMO scores correspond.
+
+    Returns:
+        tuple[PIMOResult, AUPIMOResult]: PIMO and AUPIMO results dataclass objects. See `PIMOResult` and `AUPIMOResult`.
+    """
+    if paths is not None:
+        _validate.is_source_images_paths(paths, expected_num_paths=anomaly_maps.shape[0])
+
+    # other validations are done in the numpy code
+
+    threshs, shared_fpr, per_image_tprs, _, aupimos, num_threshs_auc = functional.aupimo_scores(
+        anomaly_maps,
+        masks,
+        num_threshs,
+        fpr_bounds=fpr_bounds,
+        force=force,
+    )
+
+    pimoresult = PIMOResult(
+        threshs=threshs,
+        shared_fpr=shared_fpr,
+        per_image_tprs=per_image_tprs,
+        paths=paths,
+    )
+    aupimoresult = AUPIMOResult.from_pimoresult(
+        pimoresult,
+        fpr_bounds=fpr_bounds,
+        # not `num_threshs`!
+        # `num_threshs` is the number of thresholds used to compute the PIMO curve
+        # this is the number of thresholds used to compute the AUPIMO integral
+        num_threshs_auc=num_threshs_auc,
+        aupimos=aupimos,
+    )
+    return pimoresult, aupimoresult
